注册脚本/requirements.txt
pythontextnow>=1.1.0
requests>=2.28.0
phonenumbers>=8.12.0
pyyaml>=6.0